+++ src/entry.rs
@@ -36,7 +36,10 @@ impl Entry {
     }
 }
 
-pub fn parse(config: config::Config, line: &str) -> Result<Entry, EntryError> {
+pub fn parse(
+    config: &config::Config,
+    line: &str,
+) -> Result<Entry, EntryError> {
     let entry_pattern = Regex::new(concat!(
         r"^(?P<ws0>\s*)-(?P<ws1>\s*)\((?P<category>[a-zA-Z0-9\-]+)\)",
         r"(?P<ws2>\s*)\[(?P<bs>\\)?#(?P<pr>\d+)]",
@@ -50,13 +53,11 @@ pub fn parse(config: config::Config, line: &str) -> Result<Entry, EntryError> {
     };
 
     // NOTE: calling unwrap here is okay because we checked that the pattern matched above
-    //
-    // TODO: This should definitely improved if possible, using some iterator stuff maybe?
     let category = matches.name("category").unwrap().as_str();
     let description = matches.name("desc").unwrap().as_str();
     let link = matches.name("link").unwrap().as_str();
     let pr_number = matches.name("pr").unwrap().as_str().parse::<u16>().unwrap();
-    let spaces = vec![
+    let spaces = [
         matches.name("ws0").unwrap().as_str(),
         matches.name("ws1").unwrap().as_str(),
         matches.name("ws2").unwrap().as_str(),
@@ -65,30 +66,23 @@ pub fn parse(config: config::Config, line: &str) -> Result<Entry, EntryError> {
     ];
 
     let mut problems: Vec<String> = Vec::new();
-    for whitespace_problem in check_whitespace(spaces) {
-        problems.push(whitespace_problem)
+
+    check_whitespace(spaces)
+        .into_iter()
+        .for_each(|p| problems.push(p));
+
+    let (fixed_category, category_problems) = check_category(config, category);
+    category_problems.into_iter().for_each(|p| problems.push(p));
+
+    if matches.name("bs").is_some() {
+        problems.push("There should be no backslash in front of the # in the PR link".to_string());
     }
 
-    let (fixed_category, category_problems) = check_category(&config, category);
-    for category_problem in category_problems {
-        problems.push(category_problem)
-    }
+    let (fixed_link, link_problems) = check_link(config, link, pr_number);
+    link_problems.into_iter().for_each(|p| problems.push(p));
 
-    match matches.name("bs") {
-        Some(_) => problems
-            .push("There should be no backslash in front of the # in the PR link".to_string()),
-        _ => (),
-    }
-
-    let (fixed_link, link_problems) = check_link(&config, link, pr_number);
-    for link_problem in link_problems {
-        problems.push(link_problem)
-    }
-
-    let (fixed_desc, desc_problems) = check_description(&config, description);
-    for desc_problem in desc_problems {
-        problems.push(desc_problem)
-    }
+    let (fixed_desc, desc_problems) = check_description(config, description);
+    desc_problems.into_iter().for_each(|p| problems.push(p));
 
     let fixed = build_fixed(
         fixed_category.as_str(),
@@ -100,7 +94,6 @@ pub fn parse(config: config::Config, line: &str) -> Result<Entry, EntryError> {
     Ok(Entry {
         fixed, // TODO: why is it not possible to have this as &'a str too?
         pr_number,
-        // TODO: implement describing problems in line
         problems,
     })
 }
@@ -112,7 +105,7 @@ fn build_fixed(cat: &str, link: &str, desc: &str, pr: u16) -> String {
 
 /// Check if the category is valid and return a fixed version that addresses
 /// well-known problems.
-fn check_category(config: &config::Config, category: &str) -> (String, Vec<String>) {
+pub fn check_category(config: &config::Config, category: &str) -> (String, Vec<String>) {
     let mut problems: Vec<String> = Vec::new();
     let fixed = category.to_lowercase();
     if category.to_lowercase() != category {
@@ -130,14 +123,13 @@ fn check_category(config: &config::Config, category: &str) -> (String, Vec<Strin
 fn check_link(config: &config::Config, link: &str, pr_number: u16) -> (String, Vec<String>) {
     let mut problems: Vec<String> = Vec::new();
 
-    // TODO: check the base url of the used Git repository automatically
     let fixed = format!("{}/pull/{}", config.target_repo, pr_number);
 
     if !link.starts_with(config.target_repo.as_str()) {
         problems.push(format!("PR link points to wrong repository: {}", link))
     }
 
-    let split_link: Vec<&str> = link.split("/").collect();
+    let split_link: Vec<&str> = link.split('/').collect();
     let contained_pr_number = split_link
         .last()
         .expect("this should never be empty")
@@ -154,7 +146,7 @@ fn check_link(config: &config::Config, link: &str, pr_number: u16) -> (String, V
     (fixed, problems)
 }
 
-fn check_description(config: &config::Config, desc: &str) -> (String, Vec<String>) {
+pub fn check_description(config: &config::Config, desc: &str) -> (String, Vec<String>) {
     let mut fixed = desc.to_string();
     let mut problems: Vec<String> = Vec::new();
 
@@ -171,15 +163,13 @@ fn check_description(config: &config::Config, desc: &str) -> (String, Vec<String
         .chars()
         .last()
         .expect("no characters found in description");
-    if last_letter.to_string() != ".".to_string() {
+    if last_letter.to_string() != '.'.to_string() {
         fixed = desc.to_string() + ".";
         problems.push(format!("PR description should end with a dot: '{}'", desc))
     }
 
     let (fixed, spelling_problems) = check_spelling(config, fixed.as_str());
-    for prob in spelling_problems {
-        problems.push(prob)
-    }
+    spelling_problems.into_iter().for_each(|p| problems.push(p));
 
     (fixed, problems)
 }
@@ -222,45 +212,35 @@ fn compile_regex(pattern: &str) -> Result<Regex, Error> {
 /// Matching patterns inside of code blocks, links or within another word are ignored.
 fn get_spelling_match(pattern: &str, text: &str) -> Result<String, MatchError> {
     // Check if pattern is inside a code block
-    match RegexBuilder::new(format!(r"`[^`]*({pattern})[^`]*`").as_str())
+    if RegexBuilder::new(format!(r"`[^`]*({pattern})[^`]*`").as_str())
         .case_insensitive(true)
         .build()?
         .find(text)
+        .is_some()
     {
-        Some(_) => return Err(MatchError::MatchInCodeblock),
-        None => (),
+        return Err(MatchError::MatchInCodeblock);
     }
 
     // Check isolated words (i.e. pattern is not included in another word)
-    let found = match RegexBuilder::new(format!(r"(^|\s)({pattern})($|[\s.])").as_str())
+    match RegexBuilder::new(format!(r"(^|\s)({pattern})($|[\s.])").as_str())
         .case_insensitive(true)
         .build()?
         .captures(text)
     {
-        Some(m) => m,
-        None => return Err(MatchError::NoMatchFound),
-    };
-
-    // TODO: merge with match above to avoid double matching?
-    match found.get(2) {
-        Some(m) => Ok(m.as_str().to_string()),
-        None => return Err(MatchError::NoMatchFound),
+        Some(m) => match m.get(2) {
+            Some(m) => Ok(m.as_str().to_string()),
+            None => Err(MatchError::NoMatchFound),
+        },
+        None => Err(MatchError::NoMatchFound),
     }
 }
 
 /// Checks the used whitespace in the entry.
-fn check_whitespace(spaces: Vec<&str>) -> Vec<String> {
-    if spaces.len() != 5 {
-        panic!(
-            "unexpected amount of whitespace values passed; expected 5; got {}",
-            spaces.len()
-        )
-    }
-
+fn check_whitespace(spaces: [&str; 5]) -> Vec<String> {
     let mut problems: Vec<String> = Vec::new();
-    let separator = " ";
 
-    let errors = vec![
+    let expected_whitespace = ["", " ", " ", "", " "];
+    let errors = [
         "There should be no leading whitespace before the dash",
         "There should be exactly one space between the leading dash and the category",
         "There should be exactly one space between the category and the PR link",
@@ -268,19 +248,13 @@ fn check_whitespace(spaces: Vec<&str>) -> Vec<String> {
         "There should be exactly one space between the PR link and the description",
     ];
 
-    for (i, val) in spaces.iter().enumerate() {
-        match i {
-            // The whitespace at these indices should be empty instead of the separator
-            0 | 3 => {
-                if (*val).ne("") {
-                    problems.push(errors[i].to_string())
-                }
-            }
-            _ => {
-                if (*val).ne(separator) {
-                    problems.push(errors[i].to_string())
-                }
-            }
+    for ((got, expected), error) in spaces
+        .into_iter()
+        .zip(expected_whitespace.into_iter())
+        .zip(errors.into_iter())
+    {
+        if (*got).ne(expected) {
+            problems.push(error.to_string())
         }
     }
 
@@ -303,7 +277,7 @@ mod entry_tests {
             "- (cli) [#1](https://github.com/MalteHerrmann/changelog-utils/pull/1) ",
             "Add initial Python implementation."
         );
-        let entry_res = parse(load_test_config(), example);
+        let entry_res = parse(&load_test_config(), example);
         assert!(entry_res.is_ok());
         let entry = entry_res.unwrap();
         assert_eq!(entry.fixed, example); // NOTE: since line is okay there are no changes to it in the fixed version
@@ -315,8 +289,7 @@ mod entry_tests {
     fn test_fail_has_backslash_in_link() {
         let example =
             r"- (cli) [\#1](https://github.com/MalteHerrmann/changelog-utils/pull/1) Test.";
-        let entry_res = parse(load_test_config(), example);
-        // TODO: should this actually return an error? Not really, because parsing has worked??
+        let entry_res = parse(&load_test_config(), example);
         assert!(entry_res.is_ok());
         let entry = entry_res.unwrap();
         assert_eq!(entry.fixed, example.replace(r"\", ""));
@@ -332,7 +305,7 @@ mod entry_tests {
     fn test_fail_wrong_pr_link_and_missing_dot() {
         let example = r"- (cli) [#2](https://github.com/MalteHerrmann/changelog-utils/pull/1) Test";
         let fixed = r"- (cli) [#2](https://github.com/MalteHerrmann/changelog-utils/pull/2) Test.";
-        let entry_res = parse(load_test_config(), example);
+        let entry_res = parse(&load_test_config(), example);
         assert!(entry_res.is_ok());
         let entry = entry_res.unwrap();
         assert_eq!(entry.fixed, fixed);
@@ -354,7 +327,7 @@ mod entry_tests {
     fn test_malformed_entry() {
         let example = r"- (cli) [#13tps://github.com/Ma/2";
         // TODO: figure how to still return an entry but with the corresponding array of problems filled
-        assert!(parse(load_test_config(), example).is_err());
+        assert!(parse(&load_test_config(), example).is_err());
     }
 
     #[test]
@@ -363,7 +336,7 @@ mod entry_tests {
             r"- (cli)   [#1] (https://github.com/MalteHerrmann/changelog-utils/pull/1) Run test.";
         let expected =
             r"- (cli) [#1](https://github.com/MalteHerrmann/changelog-utils/pull/1) Run test.";
-        let entry_res = parse(load_test_config(), example);
+        let entry_res = parse(&load_test_config(), example);
         assert!(entry_res.is_ok());
         let entry = entry_res.unwrap();
         assert_eq!(entry.fixed, expected);
@@ -371,7 +344,7 @@ mod entry_tests {
         assert_eq!(entry.problems.len(), 2);
         assert_eq!(
             entry.problems,
-            vec![
+            [
                 "There should be exactly one space between the category and the PR link",
                 "There should be no whitespace inside of the markdown link",
             ]
@@ -394,14 +367,14 @@ mod category_tests {
     fn test_fail_invalid_category() {
         let (fixed, problems) = check_category(&load_test_config(), "invalid");
         assert_eq!(fixed, "invalid");
-        assert_eq!(problems, vec!["invalid change category: (invalid)"]);
+        assert_eq!(problems, ["invalid change category: (invalid)"]);
     }
 
     #[test]
     fn test_fail_non_lower_category() {
         let (fixed, problems) = check_category(&load_test_config(), "cLi");
         assert_eq!(fixed, "cli");
-        assert_eq!(problems, vec!["category should be lowercase: (cLi)"]);
+        assert_eq!(problems, ["category should be lowercase: (cLi)"]);
     }
 }
 
@@ -514,10 +487,7 @@ mod spelling_tests {
         let example = "Fix web--SdK.";
         let (fixed, problems) = check_spelling(&load_test_config(), example);
         assert_eq!(fixed, "Fix Web-SDK.");
-        assert_eq!(
-            problems,
-            vec!["'Web-SDK' should be used instead of 'web--SdK'"]
-        )
+        assert_eq!(problems, ["'Web-SDK' should be used instead of 'web--SdK'"])
     }
 
     #[test]
@@ -588,53 +558,52 @@ mod whitespace_tests {
 
     #[test]
     fn test_pass() {
-        // TODO: rather pass as &str?
-        let example_spaces = vec!["", " ", " ", "", " "];
+        let example_spaces = ["", " ", " ", "", " "];
         assert!(check_whitespace(example_spaces).is_empty());
     }
 
     #[test]
     fn test_fail_leading_space() {
-        let example_spaces = vec![" ", " ", " ", "", " "];
+        let example_spaces = [" ", " ", " ", "", " "];
         assert_eq!(
             check_whitespace(example_spaces),
-            vec!["There should be no leading whitespace before the dash"]
+            ["There should be no leading whitespace before the dash"]
         );
     }
 
     #[test]
     fn test_fail_space_between_category_and_link() {
-        let example_spaces = vec!["", " ", "", "", " "];
+        let example_spaces = ["", " ", "", "", " "];
         assert_eq!(
             check_whitespace(example_spaces),
-            vec!["There should be exactly one space between the category and the PR link"]
+            ["There should be exactly one space between the category and the PR link"]
         );
     }
 
     #[test]
     fn test_fail_multiple_spaces() {
-        let example_spaces = vec!["", "", " ", "", " "];
+        let example_spaces = ["", "", " ", "", " "];
         assert_eq!(
             check_whitespace(example_spaces),
-            vec!["There should be exactly one space between the leading dash and the category"]
+            ["There should be exactly one space between the leading dash and the category"]
         );
     }
 
     #[test]
     fn test_fail_multiple_spaces_before_description() {
-        let example_spaces = vec!["", " ", " ", "", "  "];
+        let example_spaces = ["", " ", " ", "", "  "];
         assert_eq!(
             check_whitespace(example_spaces),
-            vec!["There should be exactly one space between the PR link and the description"]
+            ["There should be exactly one space between the PR link and the description"]
         );
     }
 
     #[test]
     fn test_fail_space_in_link() {
-        let example_spaces = vec!["", " ", " ", " ", " "];
+        let example_spaces = ["", " ", " ", " ", " "];
         assert_eq!(
             check_whitespace(example_spaces),
-            vec!["There should be no whitespace inside of the markdown link"]
+            ["There should be no whitespace inside of the markdown link"]
         );
     }
 }