+++ src/init.rs
@@ -1,6 +1,5 @@
-use crate::{config::Config, errors::InitError};
-use serde_json::to_string_pretty;
-use std::{collections::HashMap, fs, path::PathBuf};
+use crate::{config::Config, errors::InitError, github::get_origin};
+use std::{collections::BTreeMap, fs, path::PathBuf};
 
 /// Runs the logic to initialize the changelog utilities
 /// in the current working directory.
@@ -18,33 +17,39 @@ pub fn init_in_folder(target: PathBuf) -> Result<(), InitError> {
     }
 
     let config_path = target.join(".clconfig.json");
-    match fs::read_to_string(config_path.clone()) {
-        Ok(_) => Err(InitError::ConfigAlreadyFound),
-        Err(_) => Ok(fs::write(config_path.as_path(), create_default_config())?),
-    }
+    let mut config = create_default_config();
+
+    if let Ok(origin) = get_origin() {
+        config.target_repo.clone_from(&origin);
+        println!("configured target repository: {}", origin);
+    };
+
+    if fs::read_to_string(&config_path).is_ok() {
+        return Err(InitError::ConfigAlreadyFound);
+    };
+
+    Ok(config.export(config_path.as_path())?)
 }
 
 /// Creates a new default configuration file for the tool.
-fn create_default_config() -> String {
-    let mut default_change_types: HashMap<String, String> = HashMap::new();
+fn create_default_config() -> Config {
+    let mut default_change_types: BTreeMap<String, String> = BTreeMap::new();
     default_change_types.insert("Bug Fixes".into(), "bug\\s*fixes".into());
     default_change_types.insert("Features".into(), "features".into());
     default_change_types.insert("Improvements".into(), "improvements".into());
 
-    let default_config = Config {
+    Config {
         categories: Vec::new(),
         change_types: default_change_types,
-        expected_spellings: HashMap::new(),
+        expected_spellings: BTreeMap::new(),
         legacy_version: None,
         target_repo: "".to_string(),
-    };
-
-    to_string_pretty(&default_config).expect("failed to print default configuration")
+    }
 }
 
 /// Creates an empty skeleton for a changelog.
 pub fn create_empty_changelog() -> String {
-    vec![
+    [
         "<!--",
         "This changelog was created using the `clu` binary",
         "(https://github.com/MalteHerrmann/changelog-utils).",