+++ src/lib.rs
@@ -6,6 +6,8 @@ pub mod cli_config;
 pub mod config;
 mod entry;
 pub mod errors;
+mod escapes;
+pub mod github;
 pub mod init;
 pub mod lint;
 mod release;